Fixed column value type and added init table instance

+++ energydb.sql
@@ -1,13 +1,13 @@
 CREATE TABLE energydb(
     created_at TIMESTAMP NOT NULL DEFAULT NOW(),
     sensor TEXT NOT NULL,
-    value FLOAT NOT NULL,
+    value NUMERIC(10,3) NOT NULL,
     PRIMARY KEY(created_at,value))
     PARTITION BY RANGE (created_at);
 
 CREATE TABLE acc_energy(
     created_at TIMESTAMP NOT NULL DEFAULT NOW(),
     sensor TEXT NOT NULL,
-    value FLOAT NOT NULL,
+    value NUMERIC(10,3) NOT NULL,
     PRIMARY KEY(created_at,value))
     PARTITION BY RANGE (created_at);
